Move version code out to new module.
