+++ src/lib.rs
@@ -7,6 +7,7 @@ mod cluster;
 mod lock;
 mod runtime;
 mod util;
+mod version;
 
 pub use cluster::{Cluster, ClusterError};
 pub use runtime::Runtime;
